requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.26.0
jinja2==3.1.3
python-multipart==0.0.6
aiosqlite==0.19.0
